VoterEquipmentController tests, no MVC tests as they are too complicated to setup
So for now I'm omitting them, since I don't have time to learn them
today (or code any abstractions to make this actually work).

Any testing is better than no-testing and right now we do have valid
unit tests for this, controllers are pretty easy to check for errors
since this covers all the major logical tests.

We should really be testing the services too, but eh, these wide tests
should cover enough cases.

A new InMemoryDAO was made explicitly so it would be possible to test
any of this.