ElectionResultsSummaryModel working + Updates (READMORE)
There's a fix to the EAVS data for provisional ballot which was not
caching correctly.

There's also an upgrade to the AutoSqlQueryable system to not have a
special case for aggregation.

This is done by just allowing multiple -arity (parameter count)
constructors and recognizing those in the construction code.

This isn't perfect, and ATM this actually ends up adding more code
(builder pattern might suck less here, but that requires more
infrastructure than I think is worth it.)

It cleans up the SqlQueryable stuff but requires more overall code.

I'll make that trade off because less magic and more explicivity matters
more than less code.