[FRONTEND] Voting Equipment Age Choropleths (#143)
* models self date

* upgrading voting equipment to be age aware.

* Age averaging logic.

* Starting to set this up...

* Refactored some stuff with helpers or moving out to other files

This is to make it easier to make the same choropleth in the front.

* Remove old testing linear gradient code

* moving styling function outside of the component to avoid more heinous hacking

If I had known how react-leaflet worked, I probably wouldn't have
written a lot of this stuff, since I obviously know a lot more about
it now than I did before, and have better ideas of how I'm actually
supposed to use any of these features.

However, it would cost more time to rearchitect as react-leaflet is
unfortunately not ultra-documented.

* gradient map legend supports position preferences

* cleanup and frontpage more progress

* fix

* choropleth in!

* new bucket colors, looks less bad

* improve readability for highlighting selectable states in splash

* FORMAT FIX

* bonus to add voting equipment by age per state county

* autoformat